refactor structure of the app
