+++ src/app/core/index.ts
@@ -4,3 +4,4 @@ export * from "./services/index";
 export { CoreModule } from "./core.module";
 export { generateBorderColor } from "./helpers/index";
 export { AuthGuard } from "./guards/index";
+export { StoreFacadeService } from "./store-facade/store-facade.service";
