+++ src/app/core/services/index.ts
@@ -3,5 +3,3 @@ export { BreadcrumbsService } from "./breadcrumbs/breadcrumbs.service";
 export { CourseService } from "./course-service/course.service";
 export { LoadingService } from "./loading-service/loading.service";
 export { ServiceModule } from "./service.module";
-export { StoreFacadeService } from "./store-facade/store-facade.service";
-
