+++ src/app/core/services/service.module.ts
@@ -3,12 +3,12 @@ import { AuthService } from "./auth-service/auth.service";
 import { BreadcrumbsService } from "./breadcrumbs/breadcrumbs.service";
 import { CourseService } from "./course-service/course.service";
 import { LoadingService } from "./loading-service/loading.service";
-import { StoreFacadeService } from "./store-facade/store-facade.service";
+
 
 @NgModule({
     declarations: [],
     imports: [],
-    providers: [AuthService, BreadcrumbsService, CourseService, LoadingService, StoreFacadeService]
+    providers: [AuthService, BreadcrumbsService, CourseService, LoadingService]
 })
 
 export class ServiceModule {}
