+++ src/app/login-page/login-page.component.ts
@@ -1,12 +1,7 @@
-import { Component, OnDestroy, OnInit } from "@angular/core";
+import { Component } from "@angular/core";
 import { Router } from "@angular/router";
-import { Store } from "@ngrx/store";
-import { Subscription } from "rxjs";
-import { map, tap } from "rxjs/operators";
 
-import * as fromApp from "../store/app.reducer";
-import * as fromAuth from "../store/reducers/auth.reducers";
-import { StoreFacadeService, IUser, IUserLogin } from "../core/index";
+import { StoreFacadeService, IUserLogin } from "../core/index";
 
 
 @Component({
@@ -14,10 +9,8 @@ import { StoreFacadeService, IUser, IUserLogin } from "../core/index";
     templateUrl: "./login-page.component.html",
     styleUrls: ["./login-page.component.scss"]
 })
-export class LoginPageComponent implements OnInit, OnDestroy {
+export class LoginPageComponent {
     private router: Router;
-    private subscription: Subscription;
-    private store: Store<fromApp.AppState>;
 
     /**
      * variable ot store user login from input
@@ -36,9 +29,8 @@ export class LoginPageComponent implements OnInit, OnDestroy {
 
     public storeFacadeService: StoreFacadeService;
 
-    constructor(router: Router, store: Store<fromApp.AppState>, storeFacadeService: StoreFacadeService) {
+    constructor(router: Router, storeFacadeService: StoreFacadeService) {
         this.router = router;
-        this.store = store;
         this.storeFacadeService = storeFacadeService;
     }
 
@@ -54,20 +46,4 @@ export class LoginPageComponent implements OnInit, OnDestroy {
         this.storeFacadeService.login(user);
     }
 
-    public ngOnInit(): void {
-
-        // this.subscription = this.store.select(fromAuth.selectAuthUserState).pipe(
-        //     tap((user) => {
-        //         console.log("ngOnint login user: ", user);
-        //         if (user) {
-        //             console.log("inside if");
-        //             this.router.navigate(["/"]);
-        //         }
-        //     })
-        // ).subscribe();
-    }
-
-    public ngOnDestroy(): void {
-        // this.subscription.unsubscribe();
-    }
 }
